refactor(editor): separate toolbar, only pass fileIndex to Editor
